+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntityApplicationModel.java
@@ -18,7 +18,6 @@
  */
 package is.codion.framework.model;
 
-import is.codion.common.state.State;
 import is.codion.common.user.User;
 import is.codion.common.version.Version;
 import is.codion.framework.db.EntityConnection;
@@ -27,7 +26,6 @@ import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.EntityType;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -46,7 +44,6 @@ public class DefaultEntityApplicationModel<M extends DefaultEntityModel<M, E, T>
 	private final EntityConnectionProvider connectionProvider;
 	private final Version version;
 	private final List<M> entityModels = new ArrayList<>();
-	private final State warnAboutUnsavedData = State.state(EntityEditModel.WARN_ABOUT_UNSAVED_DATA.get());
 
 	/**
 	 * Instantiates a new DefaultEntityApplicationModel
@@ -161,26 +158,5 @@ public class DefaultEntityApplicationModel<M extends DefaultEntityModel<M, E, T>
 		}
 
 		throw new IllegalArgumentException("EntityModel for type " + entityType + " not  found in model: " + this);
-	}
-
-	@Override
-	public final State warnAboutUnsavedData() {
-		return warnAboutUnsavedData;
-	}
-
-	@Override
-	public final boolean containsUnsavedData() {
-		return existingEntityModified(entityModels);
-	}
-
-	private static boolean existingEntityModified(Collection<? extends EntityModel<?, ?, ?>> models) {
-		for (EntityModel<?, ?, ?> model : models) {
-			EntityEditModel editModel = model.editModel();
-			if ((editModel.exists().get() && editModel.modified().get()) || existingEntityModified(model.detailModels())) {
-				return true;
-			}
-		}
-
-		return false;
 	}
 }