+++ framework/model/src/main/java/is/codion/framework/model/EntityApplicationModel.java
@@ -20,7 +20,6 @@ package is.codion.framework.model;
 
 import is.codion.common.Configuration;
 import is.codion.common.property.PropertyValue;
-import is.codion.common.state.State;
 import is.codion.common.user.User;
 import is.codion.common.version.Version;
 import is.codion.framework.db.EntityConnection;
@@ -119,18 +118,6 @@ public interface EntityApplicationModel<M extends EntityModel<M, E, T>, E extend
 	 * @return true if this model contains the given EntityModel
 	 */
 	boolean containsEntityModel(M entityModel);
-
-	/**
-	 * @return the State controlling whether this model warns about unsaved data
-	 * @see EntityEditModel#WARN_ABOUT_UNSAVED_DATA
-	 */
-	State warnAboutUnsavedData();
-
-	/**
-	 * @return true if any edit model associated with this application model contains
-	 * modified and unsaved data, that is, existing entities that have been modified but not saved
-	 */
-	boolean containsUnsavedData();
 
 	/**
 	 * @return an unmodifiable List containing the EntityModel instances contained