+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -29,6 +29,7 @@
  * {@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_SAVE_CAPTION}<br>
  * {@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_FOCUS_ACTIVATION}<br>
  * {@link is.codion.swing.framework.ui.EntityEditPanel.Config#INCLUDE_ENTITY_MENU}<br>
+ * {@link is.codion.swing.framework.ui.EntityEditPanel.Config#MODIFIED_WARNING}<br>
  * {@link is.codion.swing.framework.ui.EntityPanel.Config#USE_KEYBOARD_NAVIGATION}<br>
  * {@link is.codion.swing.framework.ui.EntityPanel.Config#DISPOSE_EDIT_DIALOG_ON_ESCAPE}<br>
  * {@link is.codion.swing.framework.ui.EntityPanel.Config#INCLUDE_TOGGLE_EDIT_PANEL_CONTROL}<br>